requirements.txt
asgiref==3.2.10
astroid==2.2.5
autopep8==1.5
beautifulsoup4==4.9.3
certifi==2020.6.20
colorama==0.4.1
Django==3.0.8
django-bootstrap4==3.0.1
django-filter==2.3.0
django-widget-tweaks==1.4.8
importlib-metadata==2.1.1
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.3.1
pytz==2020.1
soupsieve==2.2.1
sqlparse==0.3.1
typed-ast==1.4.0
wincertstore==0.2
wrapt==1.11.2
zipp==3.4.1
